+++ packages/@runlightyear/airtable/src/types/TableModel.ts
@@ -1,0 +1,11 @@
+import { Field } from "./Field";
+import { View } from "./View";
+
+export interface TableModel {
+  id: string;
+  primaryFieldId: string;
+  name: string;
+  description?: string;
+  fields: Array<Field>;
+  views: Array<View>;
+}