+++ packages/@runlightyear/airtable/src/types/WebhookRecordCreated.ts
@@ -1,0 +1,4 @@
+export interface WebhookRecordCreated {
+  createdTime: string;
+  cellValuesByFieldId: Record<string, any>;
+}