+++ packages/@runlightyear/airtable/src/types/WebhookViewChanged.ts
@@ -1,0 +1,8 @@
+import { WebhookRecordChanged } from "./WebhookRecordChanged";
+import { WebhookRecordCreated } from "./WebhookRecordCreated";
+
+export interface WebhookViewChanged {
+  changedRecordsById: Record<string, WebhookRecordChanged>;
+  createdRecordsById?: Record<string, WebhookRecordCreated>;
+  destroyedRecordIds?: string[];
+}