+++ packages/@runlightyear/lightyear/src/base/context.ts
@@ -7,9 +7,11 @@ export interface Context {
 export let context: Context = {};
 
 export function setContext(ctx: Context) {
+  console.debug("Setting context", ctx);
   context = ctx;
 }
 
 export function getContext() {
+  console.debug("Getting context", context);
   return context;
 }