+++ packages/@runlightyear/linear/src/helpers/findWorkflowStateByName.ts
@@ -2,6 +2,13 @@ import { Linear } from "../Linear";
 import { WorkflowStateResponse } from "../workflowStates/WorkflowStateResponse";
 
 export interface FindWorkflowStateByNameProps {
+  /**
+   * Linear team key. For example: "ENG"
+   */
+  teamKey: string;
+  /**
+   * Linear workflow state name. For example: "In Review"
+   */
   name: string;
 }
 
@@ -10,10 +17,10 @@ export const findWorkflowStateByName =
   async (
     props: FindWorkflowStateByNameProps
   ): Promise<WorkflowStateResponse> => {
-    const { name } = props;
+    const { teamKey, name } = props;
 
     const response = await self.listWorkflowStates({
-      filter: { name: { eq: name } },
+      filter: { team: { key: { eq: teamKey } }, name: { eq: name } },
     });
 
     return response.data.workflowStates[0];