+++ packages/@runlightyear/gsheets/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @runlightyear/gsheets
 
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies [4393656c3]
+  - @runlightyear/lightyear@0.36.0
+
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies [da7c9a679]
+- Updated dependencies [6f77253ea]
+  - @runlightyear/lightyear@0.35.0
+
 ## 0.5.0
 
 ### Minor Changes