+++ packages/@runlightyear/openai/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @runlightyear/openai
 
+## 0.6.2
+
+### Patch Changes
+
+- Updated dependencies [4393656c3]
+  - @runlightyear/lightyear@0.36.0
+
+## 0.6.1
+
+### Patch Changes
+
+- Updated dependencies [da7c9a679]
+- Updated dependencies [6f77253ea]
+  - @runlightyear/lightyear@0.35.0
+
 ## 0.6.0
 
 ### Minor Changes