+++ packages/@runlightyear/slack/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @runlightyear/slack
 
+## 0.14.6
+
+### Patch Changes
+
+- Updated dependencies [4393656c3]
+  - @runlightyear/lightyear@0.36.0
+
+## 0.14.5
+
+### Patch Changes
+
+- Updated dependencies [da7c9a679]
+- Updated dependencies [6f77253ea]
+  - @runlightyear/lightyear@0.35.0
+
 ## 0.14.4
 
 ### Patch Changes